move the server out of main pkg
